server/requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.0
flask-ipban==1.1.5
Flask-PyMongo==2.3.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
nulltype==2.3.1
plaid-python==19.0.0
pymongo==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.2
